EduID hook moved to the login helper
The EduID hook code has been moved from the login controller site to the
service helper.
The credential and reconnect controller take advantage of the service
helper EduID hook code when the session expires.
So from now on it is possible to refresh the session instead of removing
the remove the moodle instance from the site list.
At this point the integration is functional and althought the code is
not complete it works out of the box.